Update virtual-private-cloud.adoc
I believe "VPC" peering is meant here and not "VPN" peering.

According to AWS
"AWS uses the existing infrastructure of a VPC to create a VPC peering connection; it is neither a gateway nor a VPN connection, and does not rely on a separate piece of physical hardware. There is no single point of failure for communication or a bandwidth bottleneck."  https://docs.aws.amazon.com/vpc/latest/peering/what-is-vpc-peering.html

AWS explicitly states that VPC peering it is not a VPN connection.