Add note that monitoring needs to be reinstalled if agent is updated
As per this docs https://docs.mulesoft.com/release-notes/monitoring/anypoint-monitoring-installer-release-notes the monitoring installation needs to be reinstalled when upgrading the agent. So this note should be in the agent upgrade section (or both). So when you complete the upgrade of the agent, you know that you have to re-install monitoring. 

(add a link to docs on instructions on how to uninstall script and install agent)